+++ src/executor/JobExecutor.ts
@@ -2,7 +2,7 @@ import { DateTime } from 'luxon';
 
 import { ExecutionStatus, JobResult } from '../job/ExecutionInfo';
 import { ExecutionsRepository } from '../repository/ExecutionsRepository';
-import { Handler } from '../job/MomoJob';
+import { Handler, JobParameters } from '../job/MomoJob';
 import { JobEntity } from '../repository/JobEntity';
 import { JobRepository } from '../repository/JobRepository';
 import { Logger } from '../logging/Logger';
@@ -23,7 +23,7 @@ export class JobExecutor {
     this.stopped = true;
   }
 
-  async execute(jobEntity: JobEntity): Promise<JobResult> {
+  async execute(jobEntity: JobEntity, parameters?: JobParameters): Promise<JobResult> {
     const { added, running } = await this.executionsRepository.addExecution(
       this.scheduleId,
       jobEntity.name,
@@ -39,7 +39,7 @@ export class JobExecutor {
       };
     }
 
-    const { started, result } = await this.executeHandler(jobEntity);
+    const { started, result } = await this.executeHandler(jobEntity, parameters);
 
     await this.jobRepository.updateJob(jobEntity.name, {
       executionInfo: {
@@ -62,13 +62,16 @@ export class JobExecutor {
     return result;
   }
 
-  private async executeHandler(jobEntity: JobEntity): Promise<{ started: DateTime; result: JobResult }> {
+  private async executeHandler(
+    jobEntity: JobEntity,
+    parameters?: JobParameters
+  ): Promise<{ started: DateTime; result: JobResult }> {
     this.logger.debug('run job', { name: jobEntity.name });
     const started = DateTime.now();
 
     let result: JobResult;
     try {
-      const data = await this.handler();
+      const data = await this.handler(parameters);
       result = {
         status: ExecutionStatus.finished,
         handlerResult: data ?? 'finished',